pyproject.toml
[tool.poetry]
name = "ticketstest"
version = "0.1.0"
description = ""
authors = ["Aleksandr Ruban <ruban_1998@hotmail.com>"]
readme = "README.md"
packages = [{ include = "**/*.py", from = "."  }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pydantic = {extras = ["email"], version = "^2.7.1"}
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2-binary = "^2.9.9"
redis = {extras = ["hiredis"], version = "^5.0.4"}
pydantic-settings = "^2.2.1"
flask-caching = "^2.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"